Display parent artifact in the old modal
This is part of story #10738 have the list of parents when I edit a backlog item

The old modal is displayed in the content view in the AD plugin.
While creating or editing an item, it was already possible to select a
perent.

Once selected, the parent is now displayed in the "Artifact link"
section of this modale.

Change-Id: I515a69528577fd025505d940c27e8cfcd8c520b7