The burnup field can not be added twice
There is no point to have the possibility to add a burnup field twice in the
same tracker, the displayed data will be the same.

This is part of story #10670: a BurnUp chart based on Done stories

This contribution makes possible for dynamic fields to be declarated
as unique field.

Change-Id: I4a0012ba4324f8faae2055859dc94382777d29c4