request #10956: Sending a tracker notification might lead to a fatal error
If the notification is enhanced with a Message-ID header
you will get a fatal error because Zend\Mail\Message::setMessageId()
is undefined.

For some unknow reasons, while Zend\Mail\Message::setMessageId() is
a valid method according to the documentation of the ZF2 version we use [1]
the method is not present in the class delivered by the RPM package.

This issue has been introduced by commit aeb5b37a99312bfdb8e16bcd90e683a323be1fd0.

[1] https://framework.zend.com/manual/2.2/en/modules/zend.mail.additional-headers.html

Change-Id: I5bb23fff9d7102709896982f886b7f4753d7c137