request #10974: Update tools used to run the whole platform integration
To upgrade the different tools used to in the whole platform integration
tests I'm forced to switch to Chrome.
facebook/php-webdriver does not support W3C webdriver protocol ATM [1]
and I got an error when setting manually a cookie with Selenium pass through
disabled.

A small update to the tests is done because ChromeDriver takes into
account spaces when selecting a link with the text link.

[1] https://github.com/facebook/php-webdriver/issues/469

Change-Id: I22bbc29df42a74ef308e6e96a9f9dd30d6be4829