Implement Same-Site cookie and cookie prefixes protections for the Tuleap session cookie
This contribution adds two relatively new protections on the cookie used by
Tuleap to identify an authenticated user:
 * Same-Site cookie [1]: the goal of this protection is to prevent browsers to send
cookies when doing cross-site requests. It adds a new level of protection against CSRF even
if the current state of the Tuleap codebase most of the form actions can be done indifferently
with GET or POST requests. It is however a nice addition for actions done in through the REST
API. To keep an usable application, we use Same-Site cookie in the Lax mode. Also note that
using the Same-Site cookie protection helps browsers mitigating speculative side-channel
attack techniques [2]. Support by browsers is currently limited to Chrome but support
for Firefox is on the way [3].
 * Cookie prefixes [4]: the goal of this protection is to prevent cookies injection and
malicious manipulations. TL;DR: __Secure- prefix forces the cookie to be used over an
encrypted connection even if for whatever reason the Secure attribute is removed by
a malicious actor. __Host- prefix adds the same protection than the __Secure- prefix
plus it protects against subdomains setting or reading cookies set for the main
domain. Cookies prefixes are supported by Chrome and Firefox.

This is part of request #10979: Implement Same-Site cookie and cookie prefixes protections

Another contribution will add similar protection whenever possible for the cookie used
for PHP session.

Some changes are made to the tests framework to avoid forcing PHP to send data and then preventing
to test headers.

All existing users sessions are removed since the associated cookies will not be valid
anymore.

[1] https://tools.ietf.org/html/draft-west-first-party-cookies-07
[2] https://www.chromium.org/Home/chromium-security/ssca
[3] https://bugzilla.mozilla.org/show_bug.cgi?id=795346
[4] https://tools.ietf.org/html/draft-ietf-httpbis-cookie-prefixes-00

Change-Id: I7a099e96dd6bc60798c7ad2a4cba22e27be187c0