Introduce project membership delegation
This is part of story #10773: have a restricted project administrator role

Given there is a project with a static group (Eg. "Developers")
And a user is member of this static group (but is not project admin),
When you add the ugroup_id in the table project_membership_delegation
directly in the database,
Then the user see the "Admin" entry in service bar
And they can see the "members" tab when the go to project admin
And they can add (or remove) users to this project.
And they can import from a list of users
And they can (un)bind to ldap directory

Notes:
- ability to define which ugroup has membership delegation from the user
  interface will be done later

Change-Id: Ia3c4bfe059b0a9f985ab31084cfdeb6bafb27f28