request #10941: Cannot bind a ugroup to a LDAP directory group that is not at the first level of the tree
For performance reasons, the autocompleter only returns the groups found at the
first level of the tree. Groups accessible elsewhere are not in the tree
can not be selected anymore so it is not possible to bind such groups to
a Tuleap ugroup.

Due to a select2 issue [1], an empty option is added to the select tags otherwise
the change event is not trigerred the first time you use a tag not returned by the autocompleter.

[1] https://github.com/select2/select2/issues/4842

Change-Id: I6b95ee6293a5583ab2882e34cea6f4e0bafc549d