Do not try to build the SRPM of the all-deps package
The all-deps package has been removed in the commit
d7a4d9c8fd4e35068eec30539f2d18912c8a3ce1 but it has been
forgotten in the srpms target. This cause the build
system to fail if it tries to build the SRPMs.

This is part of request #10989: Do not build customization and all-deps packages

Change-Id: I25ba352a235c0bcd6a72d295a27e5414e1050f8f