Extract artifact link field to a dedicated directive and module
Part of story #10738 have the list of parents when I edit a backlog item

How to test:
- Given a tracker with a hierarchy parent, when you create a new
artifact (in TTM, create a new Bug), then you will be able to select a
parent in the list of possible parent artifacts. (Same as before)
- An error while retrieving the list of possible parent artifacts will
no longer stop the modal.
- Retrieving the list will no longer delay the modal opening.

Selecting a parent in edit will be done in later commits

Change-Id: Ie3ae83205ee8703fbbfa03509a274221742daa2c