Only project admin can see ugroups
This is part of story #10773: have a restricted project administrator role

Given a user has membership management delegation,
When the members table is displayed,
Then the "members of user groups" is omitted.

Only project administrators can see this column.
If we display the column to those that can only manage project members,
then each ugroup link will lead to an error page which is not UX
friendly.

Change-Id: Iedbe2b96e4462fc1bab17859d421959ab52d1fea