request #11004 Replace RabbitMQ queues by Redis
TL;DR it's easier to user Redis for simple queues work.
In addition to that it's easier to get a Redis up & running and it can be used
for K/V & cache store.

Worker changed a bit, I had to remove signal handling because it made the
worker script completly blind to signal (except -9). It doesn't really have
any downside as signal handling was rather poor (cannot wait for current work
to finish for instance as script is not threaded).

The code is designed to use Redis in tracker notification first with longer
goal to remove completely RabbitMQ from the dependency list.

Change-Id: I0f2e55b6abf0f1b1adb7fb95ccf505217244a5c9