request #11019: Can't access to Git repositories
The gl-membership.php script is still launched with PHP 5.3
which does not work anymore.

To test this contribution you need to redeploy gl-membership.pl
to /usr/lib/tuleap/bin/gl-membership.pl.

This is a quick fix, gl-membership.php should launched
directly through php-launcher.sh to avoid similar issue in the future.
The needed work to achieve that will be done in another contribution.

Change-Id: Ie5c911bcbafb12066946e24844268a06fb0b3c4f