Migrate Planning-v2 to BurningParrot
Part of story #11095 Style Planning-v2 in BurningParrot

Sorry for the giant review, if you think of a way to split it please
let me know.
Differences with the mockup are expected. A second patch to adjust will
be done.

How to test:
- at tuleap root, 'npm run build'

Check all actions in planning-v2:
- Top backlog vs Milestone planning
- Drag & drop
- Adding items / milestone
- Editing items / milestone
- Drop of a Task in a User Story
- Multi-drag & drop
- Filtering the backlog

Check that warnings are shown & nothing is draggable when you don't have
rights (for example not project member).

Changes:
- Removed angular-ui-router
- (Design) Replaced loader-mini by fa-spinner on Milestones updating

Change-Id: I1b6a0295d7d8a89a270a3a07037e3721569f0a21