request #11027: Changeset value can be null
If a field has no record in db on a changeset, then the method
Tracker_Artifact::getValue($field) will return null. Therefore we have
to check the returned value before calling methods on it.

I didn't manage to reproduce the case but at least we should respect the
method signature.

This contribution only targets the display of ugroup permissions in
project admin page:
Given an artifact has permission on artifact set to a specific ugroup
And the semantic title is set to a field that is not present in the last
changeset of the artifact
When you display the ugroup settings in project admin
Then there should be no fatal error.

Change-Id: I62fb57cc5420e180dc817de539ae2b1b393ea938