request #11000: Wrong cards order in a filtered Kanban
This commit fixes the following use case:

> When I filter in Kanban, the order of the cards change. E.g:
>
> - Create a TQL report "assigned_to = MYSELF()"
>
> - Add this report in the available filter in the Kanban
>
> - Assign to you some cards
>
> - Filter the kanban
>
> In my case, a top priority artifact is at the bottom of the column when
> the Kanban is filtered.
>

Since getMatchingIds does not order the result (it leaves this duty to
the table renderer, if needed) it is pretty difficult to add an order by
rank in this part of the code without breaking something or slowing
things down. Therefore the approach taken here is to order the paginated
collection. Since the collection has a limited size, it should not
impact too much performances.

Change-Id: Ied18357de7d0cdd61fc27557c43f7bedb3299442