request #11057: Irrelevant user info on a test not already run
When you create a test in a campaign (or when you create a campaign), it
is your name that is displayed in the test execution. This is
misleading, we should not display the user (and time) of the creation of
the test execution (we should display only when the test execution has
been updated).

Change-Id: Ib33eee5edcaef6bd52ab909ab7068a4026dc4b88