Move the main DAO of the pull request plugin to prepared statements and PDO
This contribution should not change anything at the functionnal level,
it is only a technical change leveraging what has been introduced by
commit 75c536e1254d78be0792f09c7c15ef6328c64135.

There is changes outside the DAO, there was a lot of boolean that were returned
by the DAO to verify if the query was successful. These return values were almost
never checked so they have been removed. With Tuleap\DB\DataAccessObject an exception
is always thrown when a query fails.

Other DAOs the pull request plugin will be dealt with in another contribution.

This is part of request #11074: Move pull request plugin to prepared statements and PDO

Change-Id: I88d725287f8d09d4c5bd0592543fbfd944154bf1