Move most of the DAOs of the pull request plugin to prepared statements and PDO
This contribution deals with most of the DAOs of the pull request plugin.
Commit 6cd3da6c23e49de955ed942e2c5b54e3808f4929 deals with the main one.
The DAO that deals with label is left untouched as it would means to also convert
parts of the Tuleap core (transactions are shared between the plugin and the core).

This contribution should not change anything at the functionnal level,
it is only a technical change leveraging what has been introduced by
commit 75c536e1254d78be0792f09c7c15ef6328c64135.

This is part of request #11074: Move pull request plugin to prepared statements and PDO

Change-Id: I3e216cc67b79a983b8c3d194954112535cb0063e