Display comment criterion in tracker classic search
This is part of story #10325 search in artifact comments from classic view

The comment criterion is always displayed in the search box.
It cannot be removed as other additional criteria.

You can now define a value and save it. It's also saved in session.
However, this value is not taken into account for now.

Change-Id: If4852a82ceb0c1ffadf479a5922e9f9e249d5e2b