Take comment criterion value into account
This is part of story #10325 search in artifact comments from classic view

The value provided in the comment criterion is now taken into
account to filter artifatcs.

This patch adds a refactoring to extract SQL part of comment
search between TQL and classic search to avoid code duplication.

/!\ There are some limitations (same the goes for TQL):

* Searches are done for words longer than 3 characters
* Some words are not taken in account because they are too common (like the, a, …)

They will be added in the documentation.

Change-Id: I7d0c5ce8bbf0d7e98bd4a8e84554b4152f14eb10