Static list field value cannot be empty
request #11089: Selectbox Static Values Empty Name Defect

Go to tracker » administration » field usage and edit a list field bound
to static values (status, severity, …). Click on a value to edit its
label and clear the text field. You should not be able to submit the
form. And if you are a Яe4l H4¢|<3r and you remove the required
attribute of the input field and submit the empty field, a warning is
displayed, and the value label is kept.

Change-Id: Ib864f668202ba3025fe2b5eb7fd4e18474c282dc