request #11131: nginx should try to redirect the user to the Tuleap virtualhost if needed
Currently when you try to access using the wrong hostname you will
get the default nginx page. It's not really frindly nor useful for the user.
This contribution try to smooth thing a little and try to redirect the user to
the hostname that has been given during the installation.

Change-Id: If4316a4b235948406443764cd5c3e812fc900b03