request #11048: Missing user status values in user details page
In admin/usergroup.php?user_id=xxx, not all user status are displayed.
Pending, validated, and validated restricted are missing.

If a user is in pending state, and the siteadmins go to this page:
1. The status displayed by the selectbox is active (which is wrong)
2. If the siteadmin doesn't notice it and submit the form, then the
user's status is switched to 'active' by error.

We cannot simply add values to the selectbox: when user's status
switches from pending to active through the regular approve pending
users page, then some actions are performed by the backend (e.g. send
emails). Those actions are not performed if we use the selectbox on the
user details page. This contribution is good enough to fix the issue
without having to spend too much time in the implementation of missing
bits.

Change-Id: I9615f3891feca59a0829ab85e3811b55ae4b7ef3