Link new artifact to parent milestone
This is part of story #10759 create a parent in planning view

Given the following plannings:

Epic -----> Release

US and Bugs --------> Sprint

And given the following hierarchy:

Release
|
-> Sprint

Epic
|
-> US

Another
|
-> Bugs

If I create an Epic from the Sprint planning, then this Epic is now
a backlog item of the corresponding Release.

If I create an Another from the Sprint planning, then this Another is
NOT linked to the corresponding Release. The behavior of the REST route PATCH
milestone/:id/content has been updated to do this. There must be no
impact on current usage of this route (DnD)

Change-Id: Idd2dcdc720e4eda73cf1d53fc3645f65c151378c