request #11158: Computed field value might not be visible in the tracker artifact view
The value of a computed field is not visible in the tracker artifact view
if the field not editable at the artifact submission and is not edited afterwards.
Said otherwise, no value is displayed if the Tracker_Artifact_ChangesetValue is null.

This issue has been introduced by commit d8049941e90be557f0bb16f2b40301aac9c94334.

Change-Id: I2bdc7b2d844bf931798205213b11fa7c0d064db0