Search against @status = OPEN()
Part of story #10702: define a query with status semantic

No change in tracker report, @status is still forbidden.
In cross-trackers, you can do the following queries:
- @status = OPEN() => it will return all open artifacts of the selected
  trackers
- @status != OPEN() => it will return all closed artifacts of the
  selected trackers
- use of logical operators: @status = OPEN() AND @title = 'rest'

Comparison against literal values, date, myself, or now are not allowed.

If one of the chosen trackers does not have the semantic status defined,
then an error is raised.

Change-Id: Ia8a07455f82bff5544600fae181d8f2501522199