request #11172 Run phpunit tests with php7
The main reason to switch to phpunit is php7 so better have them run by php7
as early as possible.

In order to do that some adjustement were needed on usage of old constructors.

The injectDependencies methods got removed as well because it's just useless
and its signature was de-facto incompatible (now an error in php7).

Also had to do some dirty fixes with globals due to recent addition of globals
check in phpunit config.

Change-Id: I5cf766cdd75bc3373f2925a3ef35838ecf16c25c