Introduce @submitted_on in xTracker TQL
This is part of story #10704 define a query with date fields

@submitted_on can be added in a query without error.
Equal (=) and not equal (!=) operators could be used.
However, they are not taken into account to filter artifacts for now.

There is also (a bit of) refactoring to differenciate
Semantic and AlwaysThereField. They are encapsulated into a common
space named "Metadata".

Change-Id: Ia22321d3a9715238c6f32798981e2e9acb9950eb