request #11174: Duplicated dynamic ugroups for legacy default template 100
As site admin, go to users groups management of the default template
(id: 100). You should not have plenty of dynamic groups.

This is due to the fact that dynamic groups are linked to the project id
100 and we search for static groups that are linked to project id 100.

The proposal is to remove display of dynamic users groups for default
template since they are not useful:
- from a template point of vue, dynamic ugroups are not taken into
  account
- from an administrative point of vue, default template is not used as a
  real project and all users that are allowed to access to this system
  project are super users.

Static users groups should be displayed though.

Change-Id: I2c5f5c36dff3d7ae8ed52f77a150706fd7b5dd38