request #11175: Remove iframe in continuous integration
By default, Jenkins disallows[1] display in iframe in order to prevent
clickjacking attacks[2]. Therefore, instead of displaying a blank iframe
that looks like broken, we should remove it instead.

This means that on Continuous Integration service homepage, there isn't
any iframe displayed when you click on a job or its builds.

In widgets, or in case of direct link to a CI items through cross
references:
- job #valid_sample_job target page should display a link to the job
  instead of an iframe
- build #123 target page should display a link to the build instead of
  an iframe

[1] https://github.com/jenkinsci/jenkins/commit/16931bd7bf7560e26ef98328b8e95e803d0e90f6
[2] https://tools.ietf.org/html/rfc7034

Change-Id: I1a15e065660927362434d785eb15ce57a48d753f