Manage list of supported widgets
Need a big refactoring because we actually had to delegate the management
of the configuration to plugins themselves.

Basically there are 2 kind of supported widgets so far:

- Core widgets
- Kanban

Each subsequent plugin will have to add a specific code to grant the import.

Part of request #11076 Project dashboards should be part of xml import.

Change-Id: I290be43ec6d61f2b5704ef3046ae6c2ba36a49fd