request #11179 Replace safe_mock by Mockery
This allow to remove some specific code that was not much used across the code
base and to simplify the mock system in Tuleap.

This also fix an issue with export tests that were throwing errors due to
ZipArchive + some missing stubs because safe_mock was not actually that safe.

Last, the introduction of OnGoingIntelligentMock for Mockery is mainly here to
help transition from SimpleTest to Mockery. It will avoid having to rewrite all
tests that depends on OnGoingIntelligentMock based on SimpleTest by providing
a convenient, mostly-drop in replacement.

But as we cannot share 100% of the interface, the methods that have no
equivalence in Mockery are just ignored so tests will crash at conversion.
Either the converter will be smart enough to propose a conversion or mere
mortals will have to dig into falling code.

Change-Id: I2ae0664ba044bcad6b4c052f5a99de56e5009158