Check that all trackers have a readable subon field
This is part of story #10704 define a query with date fields

Unused all Submitted on fields in a tracker used in cross tracker query:
you should get an error when you display the widget.

Use a Submitted on field but prevent current user to read it: you
should get an error when you display the widget.

Due to usage of PHPUnit, some legacy classes have been modified in order
to conform to their parent specification:
* Tracker_FormElement_Field_Date::getFieldDataFromRESTValueByField
* Tracker_FormElement_Field_SubmittedOn::fetchArtifactValueWithEditionFormIfEditable
* Tracker_FormElement_Field_SubmittedOn::validateFieldWithPermissionsAndRequiredStatus
* Tracker_FormElement_Field_SubmittedOn::afterCreate

Those changes are only here to be able to stub SubmittedOn fields. Only
the minimum has been done: other classes are not changed, no styling
effort, …

Change-Id: Ib5348beef86c526d60ba554382ed7b17e9e20354