Display generic error message instead of silently ignore the error
This is part of story #10704 define a query with date fields

When there is an error on the server (for example a fatal error) then
there is no rest error message and the json parse fails to parse a
potential xdebug stack trace (in dev environment). In that case we
should display an error to the end user, else we think that there is no
error and no matching artifacts (unless the dev tools are open).

Displaying this dummy error message indicates to the dev that their is
something going wrong.

Change-Id: I0bfecdcca476e583913779e8095a5df5ced61ea2