Parameters of from statements should also be unique
This is part of story #10704 define a query with date fields 

Test `@status != OPEN() or @status != OPEN()`, it should not return
empty results. It should return the same results than the query `@status
!= OPEN()`.

This is due to the fact that we were applying a array_unique on the From
statements, but not on their parameters.

Change-Id: I635feae4d53827182285b5e0131d3d32b972911d