request #11210 In milestone tracker report filter must not check in parent milestone
Given the following planning:

Epic --------------> Release
|                    |
US and Bugs -------> Sprints

If I filter in the US tracker report with "In sprint" and
if my US is directly linked to a release, the tracker report filter
must not retrieve all the release content.

It must retrieve only the sprint (and potential sub milestones) content.

Change-Id: Ib8b51698fbb976c342adf1922b563e7cd3b53dee