Add missing automated tests for TQL search
This is part of story #10323: search with TQL in attachments info

This patch adds automated tests for @comments and file field.

A dirty trick (the chown -R during the setup) is added
for /var/lib/tuleap/tracker folder to be owned by the default system user.
A next commit will be dedicated to fix properly this issue.

Change-Id: Idcfe564121e35b6a440f479b8cf4e2765b91d45a