request #11257 XML project import must not create tracker folder and content as root
When creating a project with the XML project import script,
if there is any attachment in the XML content,
then the folder /var/lib/tuleap/tracker will be owned by root and some content
will also be owned by root, which is wrong.

The whole folder with its content must be owned by the http user.

Change-Id: If20d9c4b1fdc092373d468eb630f31ebb9e90ca2