Create script to extract statistics
This is part of request #11258 Add cli script to extract stats

This script can generate the CSV file through the cli.
It must be run by either root or codendiadm.

The generated CSV will always uses "," as separator.
(there is no way for now to force the login in a script and get the
csv separator preference properly.)

Change-Id: I18e5c224b732a1331e88616a73c0105217fc2823