story #11279 Create a test project and a test user automatically
This plugin allows to create a new user with a test project with a few basic
informations related to this user.

Everything is named after the user.

You will need to deploy the sudoers file:

    install -o root -g root -m 0400 /usr/share/tuleap/plugins/create_test_env/etc/sudoers.d/tuleap_plugin_create_test_env /etc/sudoers.d/

At first REST call, a secret will be generated at

    /etc/tuleap/plugins/create_test_env/etc/creation_secret

You will need this secret to create projects

Update management of errors in import_project_xml so exception raised on parse
are reported as error on command line (exit code != 0)

Change-Id: I81413286c8260bacea91b6a1c3f83283cc76a162