Dynamic credential can be used to login on the instance
The user created in the database at the plugin installation is kept
even when the plugin is removed. This user might have done actions in Tuleap
and removing the entry can breaks the instance in multiple ways.

Since I'm a lazy contributor and the plugin does not yet contains usable
data and is not yet released this contribution does not come with a ForgeUpgrade
to handle the DB schema changes.
To be able to test it you will need to uninstall/re-install the plugin.

This is part of story #11239: generate dynamic credentials from HashiCorp Vault into a Tuleap instance

Change-Id: I3d78afbea278d66f1d5dc42c2ca6c7841612a3d5