request #11291: Proper caching of user's static ugroups
Have two projects A and B.
In A, create group UG_A, add regular user U.
In B, create group UG_B, add regular user U.

In A, create an artifact A1.
In B, create an artifact A2.

Make sure that A1 is parent of A2.

In tracker admin of A2, allow update of status field only to UG_B.

As U, go to A1 and A2. You should be able to update the status field.

NOTE: this issue is mitigated/hidden by the introduction of OpenGraph, so it should be
tested against an earlier version of Tuleap. Nevertheless the code should be fixed.

Change-Id: Ic6caaec25b1a223000ad3cbb08ddae8f7161c316