Deploy assets with RPM
This patch introduce an alternative way to deploy assets: re-use src/www/assets

Given the recent introduction of front controller, we have the opportunity to
develop plugins without having to deploy (and reload, and update) nginx config.
With this patch it goes one step further as even the css no longer need a
specific nginx config. All assets will be built and deploy in "core"
src/www/assets.

It works the same way for javascript (+ mandate the build step => hash <3)

It will require an update for images inclusions (but same way, it will be
an opportunity to include them with an hash).

This is part of story #10632 have the velocity of an artifact

Change-Id: I6a7ae84fb6942fd69fac753e47a6b1c632dcbc65