Draw velocity chart
Part of story #10633 - see velocity in agile dashboard

Please be aware that chart-layout-builder will be refactored in another
commit: We should extract legend and ticks formatting in other objects
in order to make it more generic. The goal of this commit is to draw a
chart, not to refactor 3 charts at a time.

How to test:
- In a project when semantic velocity is fully configured and where some
data are available and where a burndown and burnup graph are defined:
--> The burnup/down graphs are displayed properly (specially legends)
--> A bar chart named "velocity" is rendered.
--> One bar per sprint

Change-Id: I01da49376bbbd7689132650896e4e41b5186821e