Graph should only display 7 artifacts
this is part of story #10633 - see velocity in agile dashboard

Please note that the case in which some sprints have the same start date
is not supported. It will be taken into account if there is a actual use
case behind.

How to test:
Have more than 7 sprints in one release having different start dates
You should see only the 7 most recents + an info is shown

Change-Id: I979a19806ada7c7442f59a92c06e3354f783e1e4