Add Step Definition field
Part of story #11390: have base of steps across the stack

*Important:* This commit depends on https://gerrit.tuleap.net/11372 and
https://gerrit.tuleap.net/11373

Run forgeupgrade & clear cache.

In a tracker you can add the new field StepDefinition. This allows to define
only one step for now (you can however add steps directly in database to
display more than one step).

If there is no step in the test definition, then an empty state is displayed.
Limited information is displayed in follow-up comments.

Steps can be added at submission or at artifact edition.

The field is not displayed in artifact modal.

The field is not displayed as criteria or as column.

The field is not updatable via masschange.

No functional changes in XML project import/export.

Please note that there are inline css in the mustache file but it will be
removed in a later commit when we will replace the table by beautiful interface
(see mockups).

Later (in dedicated commits):
- tracker restriction (only for test def tracker)
- styling
- extraction of cross references on test description save

Change-Id: I41d5c432cdbf65853090d9481941c65b950d9892