request #11364 Display a lot of project members
Given a project with a lot of project members (more than 1000) with each
user member of multiple project ugroup, the project member page could not load.

If you didn't reach the timeout on your side with a lot of users,
the page must be faster to load and be displayed.

Change-Id: I7159c30453e48293cb0124680f3f0492b1a6c706