request #11363: Method arguments should map the @url definition
If in a REST resource you declare a route with a parameter, for example @url
OPTIONS {id}, then the signature of the method should also have $id parameter.
If this is not the case, then a warning is displayed, polluting the generation
of the swagger.json and forcing developers to refresh the Tuleap page that is
using REST API.

This does not occurs on production since in this case display_errors is off.

To test: run make dev-clear-cache and access to the API explorer, you don't
have red error message and available routes are displayed without having to
refresh the page.

Note: If you are using testmanagement, then apply the related patch as well, or
move plugins/testmanagement elsewhere temporarily.

Change-Id: Ide8e6c070c2d94b62615c184da8b9cb37c83dd4b