Fix fatal error when importing an Agile Dashboard template from the web UI
This impacts the "quick start" and the import from the administration

The issue has been made more visible somehow with the introduction of
38bafe2483f05a9a407b4b5ade4f9b68bc11da4a but was present since a long time
(probably since it is possible to import an Agile Dashboard template?).
When importing an tracker report table renderer with a sort Tuleap
will store things in session which does not play well with the
redirect/end of script execution. PHP does not seem to be able
to fully serialize the content of the session so we got a fatal error.
I'm not really sure why at this point PHP is not able to serialize it
but anyway Tuleap should not store things in session when doing a XML
import.

This is part of story #11189: have a smart bread crumb

Also, this contribution saves the PHP-FPM error log at the end of the
Cypress tests. Without it, debugging this kind of issue becomes too
challenging (for my taste).

Change-Id: Id2e4227e9ac2ad342f6e4ef866d0f73b4fb78a95