request #11369: Massmail feature does not set a Reply-To header to the mails
For example, when you email the members of a project through
your "My Projects" widget, the members has no way to reply to you.
Adding the Reply-To header align the behavior with what has been done in
request #11155 and request #11127.

Also note that the site administration mass emailing feature is not
impacted by this change as it uses a dedicated workflow.

Change-Id: I57e60eac408d7e39c692b8eff29d5419e11d1be9