Run simpletest with PHP 7.0
Introduce run of simpletest test suite with PHP 7.0

It's only an very first step to put in place the infrastructure and logic.
Only a couple of tests are already running with so called up to date simpletest
version (1.1.7) and php 7.0 (1180 tests).

In order to do things progressively, we generate a whitelist of tests that pass
7.0 / SimpleTest 1.1.7 and ensure they continue to pass. When one convert tests
it's supposed to run 'make simpletest-70-update-compatibility-list' and increase
the number of tests that pass.

To run tests locally you should type 'make simpletest-70'

Part of story #11385 convert tests to simpletest/php7

Change-Id: I71d1673449919d5088a1f0bddf56537fddab4011